fix: refactor header for tab and mob
