+++ src/hooks/index.ts
@@ -1,0 +1,3 @@
+import { useMediaQueries } from './useMediaQueries';
+import { useVisibility } from './useVisibility';
+export { useMediaQueries, useVisibility };
