+++ src/hooks/useMediaQueries.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useMediaQuery } from 'react-responsive';
 
-export default function useMediaQueries() {
+export function useMediaQueries() {
   const isMobileScreen = useMediaQuery({ query: '(max-width: 767.9px)' });
   const isTabletScreen = useMediaQuery({
     query: '(min-width: 768px) and (max-width: 1439.9px)',
